Add lib/asset-similarity
This absorbs asset similarity scoring and finding of closest assets, but
also adds asset merging and a function generate prioritized asset
similarity keys, which can be used for more efficient asset correlation
vs pairwise asset similarity checks.
