+++ background/assets.ts
@@ -344,79 +344,3 @@ export function assetAmountToDesiredDecimals(
 
   return fromFixedPoint(amount, decimals, desiredDecimals)
 }
-
-/**
- * Use heuristics to score two assets based on their metadata similarity. The
- * higher the score, the more likely the asset metadata refers to the same
- * asset.
- *
- * @param a - the first asset
- * @param b - the second asset
- * @return an integer score >= 0
- */
-export function scoreAssetSimilarity(a: AnyAsset, b: AnyAsset): number {
-  let score = 0
-  if (a.symbol === b.symbol) {
-    score += 1
-  }
-  if (a.name === b.name) {
-    score += 1
-  }
-  if ("decimals" in a && "decimals" in b) {
-    if (a.decimals === b.decimals) {
-      score += 1
-    } else {
-      score -= 1
-    }
-  } else if ("decimals" in a || "decimals" in b) {
-    score -= 1
-  }
-  if ("homeNetwork" in a && "homeNetwork" in b) {
-    if (
-      a.homeNetwork.name === b.homeNetwork.name &&
-      a.homeNetwork.chainID === b.homeNetwork.chainID
-    ) {
-      score += 1
-    } else {
-      score -= 1
-    }
-  } else if ("homeNetwork" in a || "homeNetwork" in b) {
-    score -= 1
-  }
-  return score
-}
-
-/**
- * Score a set of assets by similarity to a search asset, returning the most
- * similiar asset to the search asset as long as it is above a base similiarity
- * score, or null.
- *
- * @see scoreAssetSimilarity The way assets are checked for similarity.
- *
- * @param assetToFind The asset we're trying to find.
- * @param assets The array of assets in which to search for `assetToFind`.
- * @param minimumSimilarityScore The minimum similarity score to consider as a
- *        match.
- */
-export function findClosestAssetIndex(
-  assetToFind: AnyAsset,
-  assets: AnyAsset[],
-  minimumSimilarityScore = 2
-): number | undefined {
-  const [bestScore, index] = assets.reduce(
-    ([runningScore, runningScoreIndex], asset, i) => {
-      const score = scoreAssetSimilarity(assetToFind, asset)
-      if (score > runningScore) {
-        return [score, i]
-      }
-      return [runningScore, runningScoreIndex]
-    },
-    [0, -1]
-  )
-
-  if (bestScore >= minimumSimilarityScore && index >= 0) {
-    return index
-  }
-
-  return undefined
-}
